migrate fully and split
